<UserReviews /> refactor
